requirements.txt
# requirements.txt 例
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.7.1+cu121
torchaudio==2.7.1+cu121          # cu121 ビルド
asteroid==0.7.0
whisperx==3.3.1             # ← 3.3.1 にダウングレード
reazonspeech-nemo-asr==2.1.0
librosa>=0.10
soundfile>=0.12
numpy==1.26.4               # 明示固定
tqdm
python-dotenv
Cython>=3.0
PyYAML>=6.0
